+++ packages/vm/src/compatibility.rs
@@ -30,6 +30,10 @@ const SUPPORTED_IMPORTS: &[&str] = &[
     "env.db_scan",
     #[cfg(feature = "iterator")]
     "env.db_next",
+    #[cfg(feature = "iterator")]
+    "env.db_next_key",
+    #[cfg(feature = "iterator")]
+    "env.db_next_value",
 ];
 
 /// Lists all entry points we expect to be present when calling a contract.